Struct netstring::channel::Channel

pub struct Channel { /* fields omitted */ }

Methods

impl Channel

fn map<R, W, F, U>(
reader: R,
writer: W,
incoming: SyncSender<U>,
outgoing_capacity: usize,
f: F
) -> Self where
R: ReadNetstring + Send + 'static,
W: WriteNetstring + Send + 'static,
F: Fn(String) -> Option<U> + Send + 'static,
U: Send + 'static,

fn new<R, W>(
reader: R,
writer: W,
incoming: SyncSender<String>,
outgoing_capacity: usize
) -> Self where
R: ReadNetstring + Send + 'static,
W: WriteNetstring + Send + 'static,

fn send<S: Into<String>>(&self, msg: S) -> Result<(), ChannelError>

fn flush(&self) -> Result<(), ChannelError>

Flushes all pending operations

fn send_last<S: Into<String>>(self, msg: S) -> Result<(), ChannelError>

Sends a last message and consumes the channel

The writer will be flushed before the method returns.

The following example ilustrates how to exit cleanly, this requires the last message to elicit a response from the remote side:

use netstring::channel::{Channel, ChannelError};
fn terminate(channel: Channel) -> Result<(), ChannelError> {
    // send some messages
    try!(channel.send("hello"));
    try!(channel.send("world"));
    // ..drain responses
    try!(channel.send_last("good bye"));
    Ok(())
}

Safety

You must make sure all pending responses have been received or you risk them being dropped.